Completion of opl_music
XMI playing through OPL has been placed in the opl_music class. It
"almost" handles playing back multiple songs through the same opl_music
object, but right now, it's a better idea to build a new one to play a
new song. play(), pause(), and stop() should work as expected.

This is a replacement for the music playback that used to be provided by
play_opl_xmi, but with a much cleaner interface, better encapsulation,
etc.